[🔨refactor] Notice 페이지 안의 라우터에 _lib 생성하여 통신함수 구조 변경
