+++ src/app/notice/_components/NoticeBoard.tsx
@@ -1,4 +1,4 @@
-import { getNotices } from '@/lib/noticeService';
+import { getNotices } from '../_lib/noticeService';
 import { NoticeContent } from '@/lib/types/notice';
 import NoticeItem from './NoticeItem';
 
@@ -6,7 +6,7 @@ export default async function NoticeBoard() {
   const notices: NoticeContent[] = await getNotices();
 
   return (
-    <div className="space-y-4">
+    <div className="my-2 space-y-4 border border-primary">
       {notices.map((notice) => (
         <NoticeItem key={notice.id} notice={notice} />
       ))}
